Remove whitebox and deprecate corresponding verboserunner and junitfailurehacker
